+++ app/javascript/components/supplemental_application/actions.js
@@ -2,8 +2,8 @@ import { isEmpty, find, isEqual, reject } from 'lodash'
 
 import apiService from 'apiService'
 import Alerts from 'components/Alerts'
+import { isLeaseAlreadyCreated } from 'utils/applicationDetailsUtils'
 import { convertCurrency } from 'utils/form/validations'
-import { isLeaseAlreadyCreated } from 'utils/leaseUtils'
 import { performOrDefault, performInSequence } from 'utils/promiseUtils'
 import { isChanged, filterChanged } from 'utils/utils'
