--- app/javascript/context/actionCreators/application_details/leaseUiStates.js
@@ -1,3 +1,0 @@
-export const SHOW_LEASE_STATE = 'show_lease'
-export const NO_LEASE_STATE = 'no_lease'
-export const EDIT_LEASE_STATE = 'edit_lease'
